requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via
    #   flask
    #   nltk
colorama==0.4.4
    # via click
deepspeech-tflite==0.9.3
    # via deep-neural-transcriber (setup.py)
docopt==0.6.2
    # via
    #   deep-neural-transcriber (setup.py)
    #   num2words
flask==2.0.1
    # via deep-neural-transcriber (setup.py)
idna==2.10
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
joblib==1.0.1
    # via nltk
markupsafe==2.0.1
    # via jinja2
nltk==3.6.2
    # via deep-neural-transcriber (setup.py)
num2words==0.5.10
    # via deep-neural-transcriber (setup.py)
numpy==1.17.3
    # via deepspeech-tflite
pydub==0.25.1
    # via deep-neural-transcriber (setup.py)
regex==2021.4.4
    # via nltk
requests==2.25.1
    # via deep-neural-transcriber (setup.py)
toolz==0.11.1
    # via deep-neural-transcriber (setup.py)
tqdm==4.61.0
    # via
    #   deep-neural-transcriber (setup.py)
    #   nltk
urllib3==1.26.5
    # via requests
wave==0.0.2
    # via deep-neural-transcriber (setup.py)
werkzeug==2.0.1
    # via flask
